refactor: move selector resolution and auth into framework
