+++ gateway-framework/src/gateway/http/requests/mod.rs
@@ -1,3 +1,4 @@
+pub mod auth;
 mod selector;
 
 pub use selector::{RequestSelector, RequestSelectorRejection};
